updated the tests and added a new file
