+++ slitmaskgui/tests/test_mask_configurations.py
@@ -4,10 +4,10 @@ from slitmaskgui.slit_position_table import SlitDisplay
 import json
 from unittest.mock import patch, Mock
 
-# from unittest.mock import MagicMock
-# from unittest import TestCase
+""" To test the connections between the classes we will test app.py"""
 
 MASK_NAME_1 = "name"
+MASK_NAME_2 = "other_name"
 
 
 @pytest.fixture
@@ -22,14 +22,6 @@ def setup_mask_config_class(qtbot):
     return config
 
 
-@pytest.fixture
-def setup_slit_display_class(qtbot):
-    slit_display = SlitDisplay()
-    slit_display.changed_data_dict = {14:2, 15:3}
-    qtbot.addWidget(slit_display)
-    return slit_display
-
-#maybe add a make sure they are connected pytest fixture
 def initialize_configuration(test_mask_config, sample_config_data):
     with patch.object(test_mask_config.model, 'beginResetModel'), \
          patch.object(test_mask_config.model, 'endResetModel'), \
@@ -57,24 +49,21 @@ def test_update_table_to_saved(setup_mask_config_class):
     assert test_mask_config.model._data[0] == ["Saved", MASK_NAME_1]
 
 
+# def test_switching_masks()
 
-def test_data_saved_signal(setup_slit_display_class,setup_mask_config_class, sample_config_data, qtbot):
-    """ This is in test mask config because it has more to do with the mask config than the slit display (data transfer) """
-    test_mask_config = initialize_configuration(setup_mask_config_class, sample_config_data)
-    test_slit_display = setup_slit_display_class
-    test_mask_config.table = Mock()
-    test_mask_config.model = Mock()
-    test_mask_config.model.get_row_num = 1
 
-    print([w["slit_width"] for w in test_mask_config.row_to_config_dict[1]])
-    with qtbot.waitSignal(test_slit_display.data_changed) as bonker:
-        temp_data_dict = test_slit_display.changed_data_dict
-        test_slit_display.data_saved()
-    assert bonker.args == [temp_data_dict]
-    
-    assert test_slit_display.changed_data_dict == {}  # cleared
-    # assert test_mask_config.row_to_config_dict[1][15]["slit_width"] == 2  # updated
-    # test_mask_config.update_table_to_saved.assert_called_once_with(0)
+# def test_export_button()
+
+
+# def test_export_all_button()
+
+
+# def test_close_button()
+
+
+# def test_open_button()
+
+
